Split kivyutils.py into focused widget modules
`katrain/gui/kivyutils.py` was a 680-line grab bag: mixins, buttons, labels,
drop-downs, the player setup block, the collapsable panels and a few canvas
drawing helpers, in no particular order, under a name that says nothing about
what is in it. Split it by what the code is for:

* `widgets/base.py` - the background and left-click mixins everything builds on
* `widgets/buttons.py` - the `Sized*`/`AutoSized*` family, with a docstring
  explaining what each part of those names means
* `widgets/labels.py` - report table cells and the clickable/scrollable labels
* `widgets/inputs.py` - the key/value and i18n drop-downs
* `widgets/panels.py` - player setup, clock, menu items, collapsable panels
* `widgets/drawing.py` - `draw_text`/`draw_circle` and the texture caches

`widgets/__init__.py` imports them all, so `import katrain.gui.widgets` still
registers every class with Kivy's Factory for the .kv files; `__main__.py` no
longer needs a star import to do that.

`CollapsablePanel` is rewritten around what its methods actually do -
`_rebuild_tabs` (options or language changed) -> `_rebuild` (lay out header and
contents) -> `_update_height` - instead of `build_options`/`build`/`_on_size`
plus a `trigger_select(ix=None)` that meant two different things depending on
the argument. `set_state` no longer rebuilds three times per call, and the
`state` binding is no longer re-added on every rebuild.

Dead code removed: `IMETextField` (Kivy's own TextInput has handled IME
composition since 2.0, and nothing referenced this subclass), `LightLabel`,
`AnalysisToggle.tri_state` and `PlayerInfo.komi`.

No behaviour change: the panel state/height/tab transcript is identical before
and after, and the rendered screens match pixel for pixel.
