+++ katrain/gui.kv
@@ -868,7 +868,6 @@
             AnalysisToggle:
                 id: eval
                 text: i18n._("analysis:dots")
-                tri_state: True
             AnalysisToggle:
                 id: hints
                 text: i18n._("analysis:topmoves")
